Added HomeScreen and QueueApp, extraced composables from MainActivity
